refactor(create): split prompts into separate files
